+++ protocol/binary.go
@@ -21,10 +21,8 @@
 package protocol
 
 import (
-	"fmt"
 	"io"
 
-	"go.uber.org/thriftrw/internal/iface"
 	"go.uber.org/thriftrw/protocol/binary"
 	"go.uber.org/thriftrw/protocol/stream"
 	"go.uber.org/thriftrw/wire"
@@ -32,6 +30,8 @@ import (
 
 // Binary implements the Thrift Binary Protocol.
 // Binary can be cast up to EnvelopeAgnosticProtocol to support DecodeRequest.
+//
+// Deprecated: binary.Default.
 var Binary Protocol
 
 // BinaryStreamer implements a streaming version of the Thrift Binary Protocol.
@@ -101,181 +101,32 @@ var BinaryStreamer stream.Protocol
 //      an invalid request.
 var EnvelopeAgnosticBinary EnvelopeAgnosticProtocol
 
-type errUnexpectedEnvelopeType wire.EnvelopeType
-
-func (e errUnexpectedEnvelopeType) Error() string {
-	return fmt.Sprintf("unexpected envelope type: %v", wire.EnvelopeType(e))
-}
-
 func init() {
-	Binary = binaryProtocol{}
-	BinaryStreamer = binaryProtocol{}
-	EnvelopeAgnosticBinary = binaryProtocol{}
+	Binary = binary.Default
+	BinaryStreamer = binary.Default
+	EnvelopeAgnosticBinary = binaryAdapter{binary.Default}
 }
 
-type binaryProtocol struct {
-	iface.Private
-}
+// adapts binary.Protocol (which returns a binary.Responder) into
+// protocol.Protocol (which returns a protocol.Responder).
+type binaryAdapter struct{ *binary.Protocol }
 
-func (binaryProtocol) Encode(v wire.Value, w io.Writer) error {
-	writer := binary.BorrowWriter(w)
-	err := writer.WriteValue(v)
-	binary.ReturnWriter(writer)
-	return err
-}
-
-func (binaryProtocol) Decode(r io.ReaderAt, t wire.Type) (wire.Value, error) {
-	reader := binary.NewReader(r)
-	value, _, err := reader.ReadValue(t, 0)
-	return value, err
-}
-
-func (binaryProtocol) EncodeEnveloped(e wire.Envelope, w io.Writer) error {
-	writer := binary.BorrowWriter(w)
-	err := writer.WriteEnveloped(e)
-	binary.ReturnWriter(writer)
-	return err
-}
-
-func (binaryProtocol) DecodeEnveloped(r io.ReaderAt) (wire.Envelope, error) {
-	reader := binary.NewReader(r)
-	e, err := reader.ReadEnveloped()
-	return e, err
-}
-
-func (binaryProtocol) Writer(w io.Writer) stream.Writer {
-	return binary.NewStreamWriter(w)
-}
-
-func (binaryProtocol) Reader(r io.Reader) stream.Reader {
-	return binary.NewStreamReader(r)
-}
-
-// DecodeRequest specializes Decode and replaces DecodeEnveloped for the
-// specific purpose of decoding request structs that may or may not have an
-// envelope.
-// This allows a Thrift request handler to transparently accept requests
-// regardless of whether the caller submits an envelope.
-// The caller specifies the expected envelope type, one of OneWay or Unary, on
-// which the decoder asserts if the envelope is present.
-//
-// This is possible because we can distinguish an envelope from a bare request
-// struct by looking at the first byte and the length of the message.
-//
-// 1. A message of length 1 containing only 0x00 can only be an empty struct.
-// 0x00 is the type ID for STOP, indicating the end of the struct.
-//
-// 2. A message of length >1 starting with 0x00 can only be a non-strict
-// envelope (not versioned), assuming the message name is less than 16MB long.
-// In this case, the first four bytes indicate the length of the method name,
-// which is unlikely to overflow into the high byte.
-//
-// 3. A message of length >1, where the first byte is <0 can only be a strict envelope.
-// The MSB indicates that the message is versioned. Reading the first two bytes
-// and masking out the MSB indicates the version number.
-// At this time, there is only one version.
-//
-// 4. A message of length >1, where the first byte is >=0 can only be a bare
-// struct starting with that field identifier. Valid field identifiers today
-// are in the range 0x00-0x0f. There is some chance that a future version of
-// the protocol will add more field types, but it is very unlikely that the
-// field type will flow into the MSB (128 type identifiers, starting with the
-// 15 valid types today).
-func (b binaryProtocol) DecodeRequest(et wire.EnvelopeType, r io.ReaderAt) (wire.Value, Responder, error) {
-	var buf [2]byte
-
-	// If we fail to read two bytes, the only possible valid value is the empty struct.
-	if count, _ := r.ReadAt(buf[0:2], 0); count < 2 {
-		val, err := b.Decode(r, wire.TStruct)
-		return val, NoEnvelopeResponder, err
-	}
-
-	// If length > 1, 0x00 is only a valid preamble for a non-strict enveloped request.
-	if buf[0] == 0x00 {
-		e, err := b.DecodeEnveloped(r)
-		if err != nil {
-			return wire.Value{}, NoEnvelopeResponder, err
-		}
-		if e.Type != et {
-			return wire.Value{}, NoEnvelopeResponder, errUnexpectedEnvelopeType(e.Type)
-		}
-		return e.Value, &EnvelopeV0Responder{
-			Name:  e.Name,
-			SeqID: e.SeqID,
-		}, nil
-	}
-
-	// Only strict (versioned) envelopes begin with the most significant bit set.
-	// This could only be confused for a type identifier greater than 127
-	// (beyond the 15 Thrift has at time of writing), or a message name longer
-	// than 16MB.
-	if buf[0]&0x80 > 0 {
-		e, err := b.DecodeEnveloped(r)
-		if err != nil {
-			return wire.Value{}, NoEnvelopeResponder, err
-		}
-		if e.Type != et {
-			return wire.Value{}, NoEnvelopeResponder, errUnexpectedEnvelopeType(e.Type)
-		}
-		return e.Value, &EnvelopeV1Responder{
-			Name:  e.Name,
-			SeqID: e.SeqID,
-		}, nil
-	}
-
-	// All other patterns are either bare structs or invalid.
-	// We delegate to the struct decoder to distinguish invalid type
-	// identifiers, outside the 0-15 range.
-	val, err := b.Decode(r, wire.TStruct)
-	return val, NoEnvelopeResponder, err
-}
-
-// noEnvelopeResponder responds to a request without an envelope.
-type noEnvelopeResponder struct{}
-
-func (noEnvelopeResponder) EncodeResponse(v wire.Value, t wire.EnvelopeType, w io.Writer) error {
-	return Binary.Encode(v, w)
+func (b binaryAdapter) DecodeRequest(et wire.EnvelopeType, r io.ReaderAt) (wire.Value, Responder, error) {
+	v, res, err := b.Protocol.DecodeRequest(et, r)
+	return v, res, err
 }
 
 // NoEnvelopeResponder responds to a request without an envelope.
-var NoEnvelopeResponder Responder = &noEnvelopeResponder{}
+//
+// Deprecated: Don't use this directly. Use DecodeRequest.
+var NoEnvelopeResponder Responder = binary.NoEnvelopeResponder
 
 // EnvelopeV0Responder responds to requests with a non-strict (unversioned) envelope.
-type EnvelopeV0Responder struct {
-	Name  string
-	SeqID int32
-}
-
-// EncodeResponse writes the response to the writer using a non-strict
-// envelope.
-func (r EnvelopeV0Responder) EncodeResponse(v wire.Value, t wire.EnvelopeType, w io.Writer) error {
-	writer := binary.BorrowWriter(w)
-	err := writer.WriteLegacyEnveloped(wire.Envelope{
-		Name:  r.Name,
-		Type:  t,
-		SeqID: r.SeqID,
-		Value: v,
-	})
-	binary.ReturnWriter(writer)
-	return err
-}
+//
+// Deprecated: Don't use this directly. Use DecodeRequest.
+type EnvelopeV0Responder = binary.EnvelopeV0Responder
 
 // EnvelopeV1Responder responds to requests with a strict, version 1 envelope.
-type EnvelopeV1Responder struct {
-	Name  string
-	SeqID int32
-}
-
-// EncodeResponse writes the response to the writer using a strict, version 1
-// envelope.
-func (r EnvelopeV1Responder) EncodeResponse(v wire.Value, t wire.EnvelopeType, w io.Writer) error {
-	writer := binary.BorrowWriter(w)
-	err := writer.WriteEnveloped(wire.Envelope{
-		Name:  r.Name,
-		Type:  t,
-		SeqID: r.SeqID,
-		Value: v,
-	})
-	binary.ReturnWriter(writer)
-	return err
-}
+//
+// Deprecated: Don't use this directly. Use DecodeRequest.
+type EnvelopeV1Responder = binary.EnvelopeV1Responder
